move boundingbox type
