+++ src/modules/photos/components/BoundingBox/index.tsx
@@ -1,16 +1,6 @@
 import React from 'react';
 import { SquareStyle, Label } from './styled';
-
-export type BoundingBoxType = {
-  xMin: number;
-  xMax: number;
-  yMin: number;
-  yMax: number;
-  label: string;
-  scaleBorder: number;
-  width: number;
-  height: number;
-};
+import { BoundingBoxType } from './types';
 
 const BoundingBox: React.FC<BoundingBoxType> = (props) => {
   const { xMin, xMax, yMin, yMax, label, width, height, scaleBorder } = props;
