+++ src/modules/photos/components/BoundingBox/types.ts
@@ -1,0 +1,10 @@
+export type BoundingBoxType = {
+  xMin: number;
+  xMax: number;
+  yMin: number;
+  yMax: number;
+  label: string;
+  scaleBorder: number;
+  width: number;
+  height: number;
+};
